* Move most of the gphoto2 canon items into the gphoto2 base. * Remove `parse_config` and put inside `load_properties`. Remove the `properties` property so it just returns a dict. * Better parsing of output and errors from gphoto. * Remove unused `wait_for_command` as `get_command_result` was being used for that purpose. * Make the `command` self-contained so it can be overridden easily. * Add some types.
